Made los resolution changeable.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1803 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Units/COB/CobInstance.cpp
@@ -1036,7 +1036,7 @@ void CCobInstance::SetUnitVal(int val, int param)
 		break;
 	case MAX_SPEED:
 		if(unit->moveType){
-			unit->moveType->SetMaxSpeed(param/SCALE);
+			unit->moveType->SetMaxSpeed(param/(float)SCALE);
 		}
 		break;
 	}
